backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.29.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pydantic[email]==2.10.6
python-dotenv==1.0.1
google-generativeai==0.7.0
reportlab==4.2.0
python-docx==1.1.2
psycopg2-binary==2.9.10
redis==5.0.1
hiredis==3.1.0
slowapi==0.1.9
fakeredis==2.34.1
resend==0.7.0
